Add files via upload
The system allows users to send SMS text with electricity token and it automatically recharges whilst not are not even home. It also allows the user to send SMS text with codes to turn ON or OFF electrical Appliances. Finally it allows you to request electricity balance by simply sending an SMS text message.